refactor: Apply coding rule compliance
- Remove braces for single statements in if conditions
- Eliminate let usage by extracting getTargetPath utility function
- Follow coding rules specified in LLM/codingRule.md
